// src/app/api/model/transfer-bad-debt-request.ts
/**
 * CrediFlow.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TransferBadDebtRequest { 
    loanContractId?: string;
    note?: string | null;
}